+++ packages/gatsby-theme-aio/src/components/GetCredential/CredentialDetailsCard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { KeyIcon, LinkOut, MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH } from './FormFields';
+import { MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH } from './FormFields';
 import { css } from '@emotion/react';
 import { AccessToken } from './Card/AccessToken';
 import { DevConsoleLink } from './Card/DevConsoleLink';
@@ -19,6 +19,7 @@ import { ReturnScopes } from './Return/ReturnScopes';
 import { ReturnAPIKey } from './Return/ReturnAPIKey';
 import { ReturnAllowedOrigins } from './Return/ReturnAllowedOrigins';
 import { ReturnOrganizationName } from './Return/ReturnOrganizationName';
+import { KeyIcon, LinkOut } from './Icons';
 
 export const CredentialDetailsCard = ({
   credentialName,
