+++ packages/gatsby-theme-aio/src/components/GetCredential/Form/CredentialName.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { css } from "@emotion/react";
-import { AlertIcon, FormFields } from "../FormFields";
+import { FormFields } from "../FormFields";
+import { AlertIcon } from '../Icons';
 
 const CredentialName = ({ nameProps, isFormValue, formData, handleChange }) => {
